+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- `$update` arg in getParentId() at interface level
+- `populateWithValues()` method
+
+### Fixed
+- `$update` arg in `getParentId()` at interface level
 
 ## [0.0.2] - 2025-11-18